Sync files with enterprise
